# openflare-server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "OpenFlare Server 管理端与 Agent API 文档。",
        "title": "OpenFlare Server API",
        "contact": {},
        "version": "3.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/tls-certificates/": {
            "get": {
                "security": [
                    {
                        "OpenFlareTokenAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TLSCertificates"
                ],
                "summary": "List TLS certificates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "OpenFlareTokenAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TLSCertificates"
                ],
                "summary": "Create TLS certificate from PEM",
                "parameters": [
                    {
                        "description": "TLS certificate payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.TLSCertificateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/tls-certificates/apply": {
            "post": {
                "security": [
                    {
                        "OpenFlareTokenAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TLSCertificates"
                ],
                "summary": "Apply TLS certificate via ACME",
                "parameters": [
                    {
                        "description": "TLS apply payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.TLSApplyInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/tls-certificates/import-file": {
            "post": {
                "security": [
                    {
                        "OpenFlareTokenAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TLSCertificates"
                ],
                "summary": "Import TLS certificate from files",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Certificate name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Remark",
                        "name": "remark",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Certificate file",
                        "name": "cert_file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Private key file",
                        "name": "key_file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/tls-certificates/{id}": {
            "get": {
                "security": [
                    {
                        "OpenFlareTokenAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TLSCertificates"
                ],
                "summary": "Get TLS certificate detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Certificate ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/tls-certificates/{id}/content": {
            "get": {
                "security": [
                    {
                        "OpenFlareTokenAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TLSCertificates"
                ],
                "summary": "Get TLS certificate PEM content",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Certificate ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/tls-certificates/{id}/convert-acme": {
            "post": {
                "security": [
                    {
                        "OpenFlareTokenAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TLSCertificates"
                ],
                "summary": "Convert uploaded TLS certificate to ACME managed certificate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Certificate ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "TLS apply payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.TLSApplyInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/tls-certificates/{id}/delete": {
            "post": {
                "security": [
                    {
                        "OpenFlareTokenAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TLSCertificates"
                ],
                "summary": "Delete TLS certificate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Certificate ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/tls-certificates/{id}/renew": {
            "post": {
                "security": [
                    {
                        "OpenFlareTokenAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TLSCertificates"
                ],
                "summary": "Renew TLS certificate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Certificate ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/tls-certificates/{id}/update": {
            "post": {
                "security": [
                    {
                        "OpenFlareTokenAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TLSCertificates"
                ],
                "summary": "Update TLS certificate from PEM",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Certificate ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "TLS certificate payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.TLSCertificateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/tls-certificates/{id}/update-acme": {
            "post": {
                "security": [
                    {
                        "OpenFlareTokenAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TLSCertificates"
                ],
                "summary": "Update ACME TLS certificate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Certificate ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "TLS apply payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.TLSApplyInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "service.TLSApplyInput": {
            "type": "object",
            "properties": {
                "acme_account_id": {
                    "type": "integer"
                },
                "auto_renew": {
                    "type": "boolean"
                },
                "disable_cname": {
                    "type": "boolean"
                },
                "dns1": {
                    "type": "string"
                },
                "dns2": {
                    "type": "string"
                },
                "dns_account_id": {
                    "type": "integer"
                },
                "key_algorithm": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "other_domains": {
                    "type": "string"
                },
                "primary_domain": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "skip_dns": {
                    "type": "boolean"
                }
            }
        },
        "service.TLSCertificateInput": {
            "type": "object",
            "properties": {
                "cert_pem": {
                    "type": "string"
                },
                "key_pem": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AccessTokenAuth": {
            "description": "Agent API 使用节点专属 Agent Token 或全局 Discovery Token",
            "type": "apiKey",
            "name": "X-Agent-Token",
            "in": "header"
        },
        "OpenFlareTokenAuth": {
            "description": "管理端 API 使用登录后返回的用户 Token",
            "type": "apiKey",
            "name": "OPENFLARE_TOKEN",
            "in": "header"
        }
    }
}